Backend team authorisation

+++ back-end/controller/team.routes.ts
@@ -1,5 +1,6 @@
 import teamService from "../service/team.service";
 import express, { Request, Response, NextFunction } from 'express';
+import { TeamInput } from "../types";
 
 
 const teamRouter = express.Router();
@@ -137,10 +138,12 @@ teamRouter.get('/:id', async (req: Request, res: Response, next: NextFunction) =
  *                 errorMessage:
  *                   type: string
  */
-teamRouter.post('/create', async (req: Request, res: Response, next: NextFunction) => {
+teamRouter.post('/create', async (req: Request & { auth: any }, res: Response, next: NextFunction) => {
     try {
-        const team = await teamService.createTeam(req.body);
-        res.status(201).json(team);
+        const role = req.auth.role;
+        const team = <TeamInput>req.body;
+        const result = await teamService.createTeam(team, role);
+        res.status(201).json(result);
     }
     catch (error) {
         next(error);
@@ -184,9 +187,11 @@ teamRouter.post('/create', async (req: Request, res: Response, next: NextFunctio
  *                 errorMessage:
  *                   type: string
  */
-teamRouter.post('/:id/user/:userId', async (req: Request, res: Response, next: NextFunction) => {
+teamRouter.post('/:id/user/:userId', async (req: Request & { auth: any }, res: Response, next: NextFunction) => {
     try {
-        const team = await teamService.addUserToTeam(parseInt(req.params.id), parseInt(req.params.userId));
+        const currentUser = req.auth.email;
+        const role = req.auth.role;
+        const team = await teamService.addUserToTeam({teamId: parseInt(req.params.id), userId: parseInt(req.params.userId), currentUser, role });
         res.status(201).json(team);
     }
     catch (error) {
@@ -231,9 +236,11 @@ teamRouter.post('/:id/user/:userId', async (req: Request, res: Response, next: N
  *                 errorMessage:
  *                   type: string
  */
-teamRouter.delete('/:id/user/:userId', async (req: Request, res: Response, next: NextFunction) => {
+teamRouter.delete('/:id/user/:userId', async (req: Request & { auth: any }, res: Response, next: NextFunction) => {
     try {
-        const team = await teamService.removeUserFromTeam(parseInt(req.params.id), parseInt(req.params.userId));
+        const role = req.auth.role;
+        const currentUser = req.auth.email;
+        const team = await teamService.removeUserFromTeam({teamId: parseInt(req.params.id), userId: parseInt(req.params.userId), currentUser, role});
         res.status(201).json(team);
     }
     catch (error) {
